--- wiki-1.0/flows/index.yml
@@ -1,9 +1,0 @@
-!Flow
-version: '1'
-pods:
-  - name: encoder
-    uses: pods/encode.yml
-    timeout_ready: 1200000
-    read_only: true
-  - name: indexer
-    uses: pods/index.yml
