--- wiki-1.0/flows/query.yml
@@ -1,14 +1,0 @@
-!Flow
-version: '1'
-with:
-  read_only: true
-  port_expose: $JINA_PORT
-pods:
-  - name: encoder
-    uses: pods/encode.yml
-    timeout_ready: 60000
-    read_only: true
-  - name: indexer
-    uses: pods/index.yml
-    polling: all
-    uses_after: _merge_matches
