--- wiki-1.0/requirements.txt
@@ -1,4 +1,0 @@
-click==7.1.2
-transformers==4.0.1
-torch==1.7.1
-jina[scipy, http]==0.9.21
