+++ wikipedia-sentences-0.8.2/.dockerignore
@@ -1,8 +1,6 @@
-.dockerignore
 .git
 .github
-.gitignore
 data
 env
 get-data.sh
-tests
+workspace.*
