+++ wikipedia-sentences-0.8.2/.gitignore
@@ -1,3 +1,4 @@
+data
 workspace*
 env
 results
