+++ wikipedia-sentences-0.8.2/flows/query.yml
@@ -1,0 +1,18 @@
+!Flow
+with:
+  read_only: true  # better add this in the query time
+  rest_api: true
+  port_expose: $JINA_PORT
+pods:
+  encoder:
+    uses: pods/encode.yml
+    parallel: $JINA_PARALLEL
+    timeout_ready: 600000
+    read_only: true
+  doc_indexer:
+    uses: pods/doc.yml
+    shards: $JINA_SHARDS
+    separated_workspace: true
+    polling: all
+    uses_after: _merge_all
+    timeout_ready: 100000 # larger timeout as in query time will read all the data
