+++ wikipedia-sentences-0.8.2/requirements.txt
@@ -1,0 +1,3 @@
+jina[devel,torch]==0.8.2
+transformers==3.5.1
+kaggle==1.5.10
