+++ wikipedia-sentences/.dockerignore
@@ -1,6 +1,8 @@
+.dockerignore
 .git
 .github
+.gitignore
 data
 env
 get-data.sh
-workspace.*
+tests
