+++ wikipedia-sentences/.gitignore
@@ -1,4 +1,3 @@
-data
 workspace*
 env
 results
