+++ wikipedia-sentences/README.md
@@ -1,5 +1,32 @@
 # Search Wikipedia Sentences with Jina
 
+<table>
+  <tr>
+    <td>
+      Input
+    </td>
+    <td>
+      1 text file with 1 sentence per line
+    </td>
+  </tr>
+  <tr>
+    <td>
+      Output
+    </td>
+    <td>
+      <pre>top_k</pre> number of sentences that match input query
+    </td>
+  </tr>
+  <tr>
+    <td>
+      Jina version
+    </td>
+    <td>
+      0.9.22
+    </td>
+  </tr>
+</table>
+
 This is an example of using [Jina](http://www.jina.ai)'s neural search framework to search through a [selection of individual Wikipedia sentences](https://www.kaggle.com/mikeortman/wikipedia-sentences) downloaded from Kaggle. It's based on code generated by `jina hub new --type app`.
 
 ## Run in Docker
@@ -7,66 +34,82 @@ This is an example of using [Jina](http://www.jina.ai)'s neural search framework
 To test this example you can run a Docker image with 30,000 pre-indexed sentences:
 
 ```sh
-docker run -p 65481:65481 jinahub/app.app.jina-wikipedia-sentences-30k
+docker run -p 45678:45678 jinahub/app.example.wikipedia-sentences-30k:0.2.8-0.9.21
 ```
 
 You can then query by running:
 
 ```sh
-curl --request POST -d '{"top_k": 10, "mode": "search",  "data": ["text:hello world"]}' -H 'Content-Type: application/json' 'http://0.0.0.0:65481/api/search'`
+curl --request POST -d '{"top_k": 10, "mode": "search",  "data": ["text:hello world"]}' -H 'Content-Type: application/json' 'http://0.0.0.0:45678/api/search'`
 ```
 
 ## Setup
 
-1. `pip install -r requirements.txt`
-2. Set up [Kaggle](https://www.kaggle.com/docs/api#getting-started-installation-&-authentication)
-3. `sh ./get_data.sh`
-4. `export JINA_DATA_PATH='data/input.txt'`
+```sh
+pip install -r requirements.txt`
+```
 
 ## Index
 
+We'll start off by indexing a small dataset of 50 sentences (`data/toy-input.txt`) to make sure everything is working:
+
 ```sh
-python app.py index
+python app.py -t index
 ```
 
-You can set the maximum documents to index with `export MAX_DOCS=500`
+To index the [full dataset](https://www.kaggle.com/mikeortman/wikipedia-sentences) (almost 900 MB):
+
+1. Set up [Kaggle](https://www.kaggle.com/docs/api#getting-started-installation-&-authentication)
+2. Run the script: `sh ./get_data.sh`
+3. Set the input file: `export JINA_DATA_FILE='data/input.txt'`
+4. Set the number of docs to index `export JINA_MAX_DOCS=500` (or whatever number you prefer. The default is `50`)
+5. Delete the old index: `rm -rf workspace`
+6. Index your new dataset: `python app.py -t index`
 
 ## Search
 
+### With REST API
+
 ```sh
-python app.py search
+python app.py -t query_restful
 ```
 
 Then:
 
 ```sh
-curl --request POST -d '{"top_k": 10, "mode": "search",  "data": ["text:hello world"]}' -H 'Content-Type: application/json' 'http://0.0.0.0:65481/api/search'
+curl --request POST -d '{"top_k": 10, "mode": "search",  "data": ["text:hello world"]}' -H 'Content-Type: application/json' 'http://0.0.0.0:45678/api/search'
 ````
 
-Or use [Jinabox](https://jina.ai/jinabox.js/) with endpoint `http://127.0.0.1:65481/api/search`
+Or use [Jinabox](https://jina.ai/jinabox.js/) with endpoint `http://127.0.0.1:45678/api/search`
+
+### From the Terminal
+
+```sh
+python app.py -t query
+```
 
 ## Build a Docker Image
 
 This will create a Docker image with pre-indexed data and an open port for REST queries.
 
-1. Run all the steps in setup and index first. Don't run anything in the query step!
+1. Run all the steps in setup and index first. Don't run anything in the search step!
 2. If you want to [push to Jina Hub](#push-to-jina-hub) be sure to edit the `LABEL`s in `Dockerfile` to avoid clashing with other images
 3. Run `docker build -t <your_image_name> .` in the root directory of this repo
-5. Run it with `docker run -p 65481:65481 <your_image_name>`
+5. Run it with `docker run -p 45678:45678 <your_image_name>`
 6. Search using instructions from [Search](#search) above
 
 ### Image name format
 
-Please use the following name format for your Docker image, otherwise it will be rejected if you want to push it to Jina Hub. Please also see my [versioning notes](#versioning-weirdness) section before which explains my versioning workaround.
+Please use the following name format for your Docker image, otherwise it will be rejected if you want to push it to Jina Hub. 
 
 ```
-jinahub/type.kind.jina-image-name:image-jina_version
+jinahub/type.kind.image-name:image-version-jina_version
 ```
 
 For example:
 
 ```
-jinahub/app.app.jina-wikipedia-sentences-30k:0.2.3-0.9.5
+jinahub/app.example.wikipedia-sentences-30k:0.2.8-0.9.21
 ```
 
 ## Push to [Jina Hub](https://github.com/jina-ai/jina-hub)
@@ -74,20 +117,3 @@ jinahub/app.app.jina-wikipedia-sentences-30k:0.2.3-0.9.5
 1. Ensure hub is installed with `pip install jina[hub]`
 2. Run `jina hub login` and paste the code into your browser to authenticate
 3. Run `jina hub push <your_image_name>`
-
-## Notes
-
-### Changes from Default
-
-At the time of writing, `jina hub new...` creates an `encode.yml` with `max_length: 96`. I changed this to `196` which gives more accurate results (i.e. the query word actually appears in the text of the results)
-
-### Versioning Weirdness
-
-At the time of writing, the version of Jina in `requirements.txt` **doesn't** match the `jina_version` label we use in our `docker build ...` command. 
-
-Why?
-
-- I built this example with Jina 0.8.2
-- Jina Hub expects you to push with the same version you built with (i.e. it would expect me to use `jina[hub]==0.8.2` to push)
-- Pushing with Jina Hub wouldn't work for me until 0.9.5. Luckily `jina hub push ...` only cares about the Docker image, not my actual code (I ran `jina[hub]==0.9.5` in a separate virtualenv)
-- We're working on updating this example code to 0.9.5 to get around this ugly kluge and delete this note!
