+++ wikipedia-sentences/app.py
@@ -1,59 +1,103 @@
-__version__ = '0.0.1'
+__copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
+__license__ = "Apache-2.0"
 
 import os
-import sys
+import itertools as it
 
+import click
 from jina.flow import Flow
+from jina import Document
 
-num_docs = int(os.environ.get('MAX_DOCS', 500))
 
 def config():
-    parallel = 2 if sys.argv[1] == 'index' else 1
-    shards = 2
+    os.environ["JINA_DATA_FILE"] = os.environ.get(
+        "JINA_DATA_FILE", "data/toy-input.txt"
+    )
+    os.environ["JINA_WORKSPACE"] = os.environ.get("JINA_WORKSPACE", "workspace")
 
-    os.environ['JINA_PARALLEL'] = str(parallel)
-    os.environ['JINA_SHARDS'] = str(shards)
-    os.environ['WORKDIR'] = './workspace'
-    os.makedirs(os.environ['WORKDIR'], exist_ok=True)
-    os.environ['JINA_PORT'] = os.environ.get('JINA_PORT', str(65481))
+    os.environ["JINA_PORT"] = os.environ.get("JINA_PORT", str(45678))
+    # os.environ["JINA_MAX_DOCS"] = os.environ.get("JINA_MAX_DOCS", str(100))
 
 
-# for index
-def index():
-    f = Flow.load_config('flows/index.yml')
+def print_topk(resp, sentence):
+    for d in resp.search.docs:
+        print(f"Ta-Dah🔮, here are what we found for: {sentence}")
+        for idx, match in enumerate(d.matches):
+
+            score = match.score.value
+            if score < 0.0:
+                continue
+            print(f'> {idx:>2d}({score:.2f}). {match.text}')
+
+
+def index(num_docs):
+    f = Flow().load_config("flows/index.yml")
 
     with f:
-        data_path = os.path.join(os.path.dirname(__file__), os.environ.get('JINA_DATA_PATH', None))
+        data_path = os.path.join(os.path.dirname(__file__), os.environ.get('JINA_DATA_FILE', None))
         f.index_lines(filepath=data_path, batch_size=16, read_mode='r', size=num_docs)
 
-# for search
-def search():
-    f = Flow.load_config('flows/query.yml')
 
+def query(top_k):
+    f = Flow().load_config("flows/query.yml")
+    with f:
+        while True:
+            text = input("please type a sentence: ")
+            if not text:
+                break
+
+            def ppr(x):
+                print_topk(x, text)
+
+            f.search_lines(lines=[text, ], on_done=ppr, top_k=top_k)
+
+
+def query_restful():
+    f = Flow().load_config("flows/query.yml")
+    f.use_rest_gateway()
     with f:
         f.block()
 
 
-# for test before put into docker
 def dryrun():
-    f = Flow.load_config('flows/query.yml')
-
+    f = Flow().load_config("flows/index.yml")
     with f:
-        pass
+        f.dry_run()
 
 
-if __name__ == '__main__':
-    if len(sys.argv) < 2:
-        print('choose between "index/search/dryrun" mode')
-        exit(1)
-    if sys.argv[1] == 'index':
-        config()
-        index()
-    elif sys.argv[1] == 'search':
-        config()
-        search()
-    elif sys.argv[1] == 'dryrun':
-        config()
+max_docs = int(os.environ.get("JINA_MAX_DOCS", 50))
+@click.command()
+@click.option(
+    "--task",
+    "-t",
+    type=click.Choice(
+        ["index", "query", "query_restful", "dryrun"], case_sensitive=False
+    ),
+)
+@click.option("--num_docs", "-n", default=max_docs)
+@click.option("--top_k", "-k", default=5)
+def main(task, num_docs, top_k):
+    config()
+    workspace = os.environ["JINA_WORKSPACE"]
+    if task == "index":
+        if os.path.exists(workspace):
+            print(f'\n +----------------------------------------------------------------------------------+ \
+                    \n |                                   🤖🤖🤖                                         | \
+                    \n | The directory {workspace} already exists. Please remove it before indexing again.  | \
+                    \n |                                   🤖🤖🤖                                         | \
+                    \n +----------------------------------------------------------------------------------+')
+        index(num_docs)
+    if task == "query":
+        if not os.path.exists(workspace):
+            print(f"The directory {workspace} does not exist. Please index first via `python app.py -t index`")
+        query(top_k)
+    if task == "query_restful":
+        if not os.path.exists(workspace):
+            print(f"The directory {workspace} does not exist. Please index first via `python app.py -t index`")
+        query_restful()
+    if task == "dryrun":
         dryrun()
-    else:
-        raise NotImplementedError(f'unsupported mode {sys.argv[1]}')
+
+
+if __name__ == "__main__":
+    main()
