+++ wikipedia-sentences/data/toy-input.txt
