+++ wikipedia-sentences/flows/index.yml
@@ -1,11 +1,9 @@
 !Flow
+version: '1'
 pods:
-  encoder:
+  - name: encoder
     uses: pods/encode.yml
-    parallel: $JINA_PARALLEL
-    timeout_ready: 600000
+    timeout_ready: 1200000
     read_only: true
-  doc_indexer:
-    uses: pods/doc.yml
-    shards: $JINA_SHARDS
-    separated_workspace: true
+  - name: indexer
+    uses: pods/index.yml
