+++ wikipedia-sentences/manifest.yml
@@ -5,7 +5,7 @@ author: Alex Cureton-Griffiths (alex.cg@jina.ai)
 url: https://github.com/alexcg1/jina-wikipedia-sentences
 vendor: Jina AI Limited
 documentation: https://github.com/alexcg1/jina-wikipedia-sentences
-version: 0.2.4
+version: 0.2.8
 license: apache-2.0
 keywords: [NLP, wikipedia, text, distilbert, example, transformers]
 type: app
