+++ wikipedia-sentences/requirements.txt
@@ -1,3 +1,4 @@
-jina[devel,torch]==0.8.2
-transformers==3.5.1
-kaggle==1.5.10
+click==7.1.2
+transformers==4.0.1
+torch==1.7.1
+jina[scipy, http]==0.9.21
